requirements.txt
numpy==1.26.4
openai-whisper==20240930
torch==2.3.0
pandas==2.2.3
tensorflow==2.18.0
keras==3.6.0
gymnasium==1.0.0
stable_baselines3==2.4.0
levenshtein==0.26.1
matplotlib==3.9.2
scipy==1.14.1
pydub==0.25.1
seaborn==0.13.2
scikit-learn==1.5.2
scikeras==0.13.0